fix cover tool and add some basic tests
Bits of this were pretty broken and hardcoded to work with the example
fizzbuzz program shown during the lightning talk.

This should now pretty work As Advertised. Very lazy script tests are
added as some sort of proof.

A new test case documents the way that this breaks if the os package is
not imported by the target somehow.