handle missing dependencies of the vars package
This change updates link and compile so that we don't depend on os (and
now bufio) already being imported by the target build. Previously, we
would detect when os was missing from the linker's importcfg and error
out.

Now, we patch importcfgs for each package while building the generated
code so that existing packages are reused. This avoids fingerprint
mismatches.

Since we don't need to be as minimal with the dependencies of the
generated package anymore, bufio is a nicer way to write out a
potentially large blob of lines to the cover file.